requirements.txt
raspberrypi-tm1637==1.3.2
requests==2.21.0
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
wiringpi==2.60.0
timeout-decorator